Refactor of FlightForm
