+++ frontend/src/pages/AddFlightPage/AddFlightPage.tsx
@@ -1,5 +1,5 @@
 import './AddFlightPage.css';
-import FlightForm from "./components/FlightForm.tsx";
+import FlightForm from "../../components/FlightForm/FlightForm.tsx";
 import {NewFlight} from "../../types/model/dataType.ts";
 import {FormEvent, useState} from "react";
 import axios from "axios";
